requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
pymupdf>=1.23.0
matplotlib>=3.7.0
Pillow>=9.5.0
numpy>=1.24.0
